Split parts of `http/mod.rs` into `http/handlers.rs`
Smaller and more focused files.
